Fix transactions applying to every turn
Transactions in the form of a number such as Quick Supply were
applying to every turn. When they did not match the current turn they
were tesed as an expression which assumes the format
{operator}{number}. Then that expression is executed and if the result
is truthy then the transaction is applied. For a simple number the
expression became {currentTurn}{turnAdded} which evaluates to a large
whole number which is truthy.

The fix is to bail out if you notice you do not have an n in the
expression, since then it was just a value, not an expression.